requirements.txt
Django==1.8.16
argparse==1.2.1
django-markdown-deux==1.0.5
markdown2==2.3.0
pytz==2015.4
six==1.9.0
wsgiref==0.1.2
djangorestframework==3.5.3
requests
